backend/requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
openapi-spec-validator 
pyyaml
dotenv
fastapi 
uvicorn
psycopg[binary,pool]
langgraph-checkpoint-postgres
sqlalchemy
psycopg2-binary
python-dotenv